Replace usage of deprecated PHPUnit `returnCallback()` function (#2085)
This PR is meant to serve as a first step for addressing changes that
will need to be made before https://github.com/Kitware/CDash/pull/1866
can be merged. The PHPUnit function `returnCallback()` is
[deprecated](https://github.com/sebastianbergmann/phpunit/issues/5423),
so our usage of it was switched to its shorthand syntax, as recommended
by the docs.