Add GITHUB_ALWAYS_PASS configuration option (#2091)
When this environment variable is set to true, CDash will always report
a status of 'success' to any GitHub repositories that have installed
this CDash server as a GitHub App.

This is useful in cases where you do not want the CDash GitHub check to
prevent a PR from being merged.

---------

Co-authored-by: William Allen <william.allen@kitware.com>