Add project visibility field to GraphQL schema (#2101)
This PR exposes a visibility field in the project type in the new
GraphQL API.

I intend to gradually phase in each of the columns in the project table
after carefully considering the name, necessary constraints, and purpose
of each one. This is the first of several planned PRs on this topic.