Add new `MAX_PROJECT_VISIBILITY` config option (#2115)
Some CDash administrators may want users to be able to create projects
on their own, but not public or protected projects. This PR adds a new
`MAX_PROJECT_VISIBILITY` environment variable which specifies the most
visible setting a project can be configured to use. Instance
administrators can override this setting, to allow some projects to be
made public or protected with administrator approval.