Add additional fields to GraphQL Build type (#2098)
The new GraphQL API is the highlight feature in the upcoming CDash 3.4
release, but we don't currently expose much information via the new API.
This PR fills in some of the gaps by adding GraphQL fields for most of
the columns in the `build` table.

Although this PR is straightforward from a functionality perspective, I
am looking for in-depth feedback during the code review process,
particularly related to the field names, types, and documentation. In
some cases, the documentation I added is sparse because I was not able
to determine the precise purpose of some columns.

---------

Co-authored-by: Zack Galbreath <zack.galbreath@kitware.com>