Add SiteInformation type to GraphQL API (#2181)
The current network of site-related CDash pages is filled with
inconsistencies, and at times, incorrect behavior, and is in dire need
of a refactor. I plan to use an overhaul of these site pages as a
testing ground for the new GraphQL API in the near future. To support
that upcoming work, this PR exposes SiteInformation data via the GraphQL
API and cleans up a bunch of related logic.