Drop unused build log column (#2145)
The log column in the build table appears to be unused. Information can
be written to the column, but there is no way to extract the contents.
Since this appears to be a legacy feature which is obsolete in CDash
today, I have removed the log column and all remaining references to it.
Build.xml files containing the LOG element will be ignored.

While I've done a significant amount of digging to try to find
additional references to the log column, it is theoretically possible
that the column is still referenced in a SQL query or via reflection
somewhere.