Move test name column to `build2test` table (#2148)
The database schema currently has a `build2test` table which contains
one row for each instance of a test run, in addition to a `test` table
which contains deduplicated test names. The joins required when
searching for `build2test` rows by test name, or for other relations by
test name through the `build2test` relation, incur a significant amount
of indexing overhead, and lead to poor query plans in some cases. The
presence of a project ID column on the test relation also encourages
poor query design, by applying the project filter to the test relation
instead of the build relation.

This PR moves the test name to the `build2test` table, and drops the
`test` table. This change is almost entirely internal, but has the
potential to boost query performance in several places, as well as
simplify future additions to the GraphQL API. The database migration
associated with this change is large and may take a significant amount
of time to apply on large CDash systems.