Remove unnecessary mix build after test (#2215)
As of https://github.com/Kitware/CDash/pull/2153, the `APP_URL`
environment variable is no longer built into the website code, and it is
no longer necessary to run a Mix build after the test suite to restore
the original `.env`. This change will save a few seconds on every CI
run.