Add GraphQL query filtering capability (#2212)
This PR introduces the ability to add conditions to aggregate queries in
the GraphQL API. This allows users to select only public projects or all
the builds submitted after a given timestamp, for example. This
introductory PR adds the =, !=, <, and > operators. Future PRs will add
additional operators to the filter set.

How is this PR different from Lighthouse-PHP's
[existing](https://lighthouse-php.com/master/eloquent/complex-where-conditions.html)
filtering functionality? The approach Lighthouse takes is not type safe,
which makes the API confusing to use and difficult to maintain. It also
lacks the flexibility CDash needs, and does not support enumeration
filtering. This PR builds upon Lighthouse's existing approach, by
introducing type-safe filters.

In the future, these filters can be improved by restricting the
operators available on a per-type basis. For example, it doesn't make
sense to perform a less-than operation on an enumeration in most cases,
but we still offer the less-than operator.