Switch legacy DB to Laravel DB facade in XML handlers (#2229)
This PR is a small amount of incremental progress towards fully
switching from the legacy database connection to Laravel's DB facade, or
Eloquent, as appropriate. This PR specifically converts all legacy
database code in the XML handlers. Changes like this lay the groundwork
for the eventual ability to log the queries being executed for every
request.