Return TZ with DateTime GraphQL results (#2225)
The GraphQL API currently implicitly returns DateTime information in
UTC, but this is not necessarily clear to end users, and may not be true
in all cases in the future. This PR adds time zone information to each
of the DateTime values returned by the API. in accordance with ISO 8601.
In the future, we should convert all of our DateTime database columns to
DateTimeTz columns.