Add initial support for `Build.xml` submission file validation (#2277)
The current error handling and logging for XML files submission parsing
is not extensive enough, and it is not easily accessible to users making
the submissions. This PR is the first in a series of changes meant to
improve this aspect of the user experience.

The changes define an initial schema for valid "Build" XML files, and a
tool to validate such files against it. The schema has been tested
against all build XML data files in the CDash repo, as well as on
production data (generated by CTest).

Future PRs are to expand on this by improving the rigor of the schema,
defining schemas for all other upload types CDash accepts (update,
configure, test, etc.), and incorporating the validation into our
submission handling process.

---------

Co-authored-by: Zack Galbreath <zack.galbreath@kitware.com>