Refactor common element types to resuable schema component (#2291)
As part of our efforts to improve the CDash submission process, we
working to introduce XML schematics to help with submission validation.
The first of these PRs merged is
https://github.com/Kitware/CDash/pull/2277. The changes in this PR
refactor some common element types that appear in many of the different
XML types, and pull them out to a common file so they can be reused and
instead of redefined by the other schema files. For an example where
these elements type are used, see
https://github.com/Kitware/CDash/pull/2287.