Move contents of `upgrade_functions.php` to controller (#2283)
All of the remaining references to `upgrade_functions.php` are in a
single controller, thanks to a number of previous PRs removing former
usages of the functions contained in this file. Moving these functions
into the controller where they are used allows us to eliminate a few of
the remaining non-autoloaded imports.