Add schema for the `Update` XML submission type (#2298)
This PR is part of a series meant to improve the submission validation
in CDash. The changes introduce an initial schema for the "Update" XML
file type accepted by the CDash submission process. As in the other PRs,
this schema has been tested against all such existing XML data files in
the CDash repo.