Add schema for the `Notes` XML submission type (#2293)
This PR is part of a series meant to improve the submission validation
in CDash. The changes introduce an initial schema for the "Notes" XML
file types accepted by the CDash submission process. As in the other
PRs, this schema has been tested against all such existing XML data
files in the CDash repo.