Change how Composer is installed for UBI-based image (#2348)
The CDash UBI-based image currently fails to start with the following
error:
```
(13)Permission denied: AH00058: Error retrieving pid file run/httpd.pid
AH00059: Remove it before continuing if it is corrupted.
```

This error is coming from an attempt to set up the prerequisites for
Composer to be installed, which is inadvertently breaking httpd.

This PR resolves the issue by installing Composer the preferred way, by
downloading it directly from the Composer website.