Add basic Test type to GraphQL API (#2346)
Our current representation of test results is complex, and will take a
significant amount of work to expose via the GraphQL API safely. This PR
starts that process by exposing test names and statuses via the GraphQL
API. In addition, this PR renames the `BuildTest` model to `Test`. The
underlying table will be renamed at a future point.