Refactor legacy UserProject model (#2396)
The legacy UserProject model will eventually be superseded by Eloquent
relationships between the User and Project models. This PR contains
incremental progress towards the goal of getting rid of this legacy
model. I expect to finish this refactor in a separate PR once a handful
of other changes are complete. For now, most of the "easy" work is done.