Use Eloquent Project model for test project creation (#2399)
The functions used to create projects for much of the test suite
currently return legacy project models, even though most tests
immediately convert them to Eloquent models. This make the test-writing
process extra painful for most tests, and is completely unnecessary.

This PR changes the project creation functions to instead return
Eloquent projects. This also removes some of the last references to the
InitialSetup method in the legacy project model, which paves the way for
that method to be completely removed in the follow-up to #2396.