Add new filters UI and barebones viewTest.php replacement (#2439)
The current filters interface is written in AngularJS, and the backend
for the filters interface currently suffers from a variety of
performance, security, and maintainability issues. The new GraphQL API
was designed to replace the legacy JSON API, including the existing
filtering logic. This PR introduces a new filters interface which uses
GraphQL introspection to automatically determine the available fields,
operators, and enum values (if applicable) for a given filter type.
Since the filters interface accepts an initial set of GraphQL filters
and returns modified GraphQL filters, it's completely reusable and
straightforward to add to new pages. This PR is also the first time
DaisyUI has been used for a page in CDash.

To provide a basic use case for the new filters component, I created a
skeleton replacement for viewTest.php which displays test names and
statuses via a paginated GraphQL query. This page should appear to load
much faster than the existing viewTest.php due to the pagination.

Although this PR introduces a significant amount of new functionality,
there is still much more to do, including:
- I intend to create a follow-up PR which builds upon our existing
Cypress component testing infrastructure, after which thorough tests can
be written for the individual components which make GraphQL queries.
- The creation of a `cdash-link` CSS class was needed in this PR to
prevent conflicts between DaisyUI and existing CDash CSS code. This work
should be continued to reduce the number of conflicts between DaisyUI
and CDash CSS.
- The Boolean and DateTime filter types need to be implemented in the
UI. This requires a date picker to be added, and will be done in a
separate PR.
- More columns will be added to the new tests page when the associated
data becomes available in the GraphQL API.

<hr>


![image](https://github.com/user-attachments/assets/f00c83b3-5ae6-486b-86a7-49d8ad15f94f)