Improve `/projects` load performance (#2454)
`/projects` currently shows a loading indicator briefly on large
production instances with many builds. This PR attempts to address the
issue by adding a `buildCount` field to the Project type in the GraphQL
API, which allows filtered count queries to be executed more
efficiently. I plan to add similar fields for build errors/warnings,
tests, etc in the future.