Create storage dirs before enabling maintenance mode (#2459)
Laravel's maintenance mode writes a file to the storage directory, which
is problematic for new instances which do not have storage directories
created yet. This PR moves the directory creation logic before the
command to turn on maintenance mode, which allows new instances to be
spun up seamlessly without needing commands to be run manually.