Allow session cookies over HTTP if APP_URL starts with http:// (#2462)
https://github.com/Kitware/CDash/pull/2458 removed the requirement for
production sites to operate over HTTPS, but did not change the
configuration option which allows session cookies to be sent over HTTP.
This PR changes that option to allow session cookies to work over HTTP.
Operating a CDash site over HTTP is insecure and is not recommended for
most users.

User sessions don't really make sense for a HTTP site, since sessions
cannot be secured. If there is interest and financial support in the
future, it may be interesting to investigate simply turning off the
ability to log in altogether for HTTP sites.

Closes #2461