Make autoremove batch size configurable
We've noticed some large production sites struggle to delete large
batches of builds in a performant manner.

To help address this issue, this commit introduces a new configuration option
to control the number of builds that CDash attempts to remove in a single pass.

This commit also reduces the default batch size from 100 to 10.