Run worker containers with unlimited memory (#2490)
Override the memory_limit from php.ini (currently hardcoded to 512M) for
worker containers.

The reasoning here is that the `php artisan queue:work` command is
running by itself in the container, so it should be allowed to consume
as much memory as necessary.

Memory requests/limits should be specified at the Kubernetes or docker
compose level instead.