Overhaul footer layout and add GraphQL Explorer link (#2498)
The footer currently has a number of formatting issues on various page
styles due to our inconsistent use of styling and web frameworks. This
PR aims to clean up some of the inconsistencies. Additionally, this PR
adds a new "GraphQL Explorer" footer link and removes the "View as JSON"
link on pages which do not have a JSON API to display.


![image](https://github.com/user-attachments/assets/cc38ee7c-ddb8-4548-b544-e18e06de620b)