Move buildinformation table content to build table (#2499)
The `buildinformation` table currently has a one-to-one relationship
with the `build` table, which provides no real benefits, and makes
querying the build table based on a combination of build and
buildinformation data inefficient. This PR moves the contents of
`buildinformation` to the `build` table. This PR also exposes the build
information data as filterable fields in the GraphQL API.