Fix blank lat/long GraphQL Site fields (#2504)
Fixes #2502. The type for the latitude and longitude fields is a string
in the database but a float in the API, which causes failures when empty
strings are present. This PR converts the API type to a string to
resolve the issue. In the future, it may be worth creating a custom
scalar type for lat/long information.