Force URL to start with HTTPS if APP_URL starts with https:// (#2508)
https://github.com/Kitware/CDash/pull/2458 removed a line which was
believed to have no effect. After deploying a recent version of CDash to
a production system based on Kubernetes, it was discovered that Laravel
automatically changes the protocol to `http://` if the incoming request
is a HTTP request, even if the `APP_URL` starts with `https://`. This PR
addresses the issue by forcing Laravel to respect the protocol specified
by the `APP_URL` environment variable.