Add Coverage type to GraphQL API (#2515)
This PR continues our ongoing effort to expose major components of the
CDash database via the GraphQL API. As of now, these changes are rather
useless since the filename is not stored in the Coverage table. A future
PR will migrate the filename to the underlying coverage table, and add a
filename field to the Coverage GraphQL type.