Overhaul project sites page (#2583)
The "site map" page is currently very slow on large production
instances, and is also partially broken. This PR completely overhauls
the page, using Vue and GraphQL behind the scenes to render content
faster. I plan to make a series of UI improvements to this page in the
near future via separate PRs.


![image](https://github.com/user-attachments/assets/1560f9d8-2c26-41f6-a520-2639b41d8a7a)