Add submission trait to test environment (#2578)
There is currently no way to create submissions programmatically in the
new part of the test suite. Given that various upcoming features will
need test submissions to be made, this PR moves a preexisting database
seeder method to a test trait where it can be used more widely. I also
replaced the legacy `submission()` method which was previously the
preferred way to make submissions in the old part of the test suite.