Replace cdashmail() email logging with Laravel event listener (#2623)
The preferred way to handle email notifications in Laravel is to use
[Mailables](https://laravel.com/docs/11.x/mail#writing-mailables). Any
emails sent outside the legacy `cdashmail()` function, including
Mailables, cannot take advantage of the email logging functionality
`cdashmail()` provides.

This PR moves that logging functionality to an "Event
Listener"--Laravel's preferred way to respond to events such as emails
being sent. `cdashmail()` was also removed since the logging changes
rendered it obsolete.

I plan to make a series of PRs to convert each of our raw text emails to
Mailables in the future.