Remove `cdash_unlink()` helper (#2650)
The `cdash_unlink()` helper function is only used in a handful of
locations, and most of those locations do not appear to benefit from its
behavior. This PR removes the `cdash_unlink()` helper and replaces it
with either the raw PHP `unlink()` command or Laravel's Storage facade's
delete helper. In the future, I plan to create a dedicated Laravel
temporary disk which will allow us to use the storage facade for all
file I/O operations.