Add support for validation of XML submission files (#2437)
This is a follow up to the submission XML schemas along with the
validation artisan command (see
https://github.com/Kitware/CDash/pull/2335) that were recently
introduced. The changes in this PR add the option to use this validation
process on incoming submission files, although this option is disabled
by default.
A test has been added to verify these changes, and the docs have been
updated to reflect the new configuration option.

---------

Co-authored-by: Joseph Snyder <joe.snyder@kitware.com>
Co-authored-by: William Allen <16820599+williamjallen@users.noreply.github.com>