Overhaul site information page (#2661)
The site information page `/sites/<id>` is currently somewhat broken,
and does not present much useful information. This PR is the first of
several planned as part of a revitalization of the CDash site feature.
The entire UI was rebuilt with DaisyUI based on information provided via
the GraphQL API. Further improvements remain to be made, including:
- The "claim site" feature should be moved to this page
- Site information should be editable directly from this page
- Aggregate time per project, current site location, and build status
information should be added

Closes https://github.com/Kitware/CDash/issues/1126

Related to https://github.com/Kitware/CDash/issues/1984

New UI:

![image](https://github.com/user-attachments/assets/395d7883-469a-470f-a538-b73a4d3a616d)