Start using autoregister as the default for OAuth (#2677)
Remove the configuration options for the OAuth providers which would
allow for the auto-registration of a user to be disabled.

Remove the code paths that would allow either the user to fill the
registration form for themselves or the error that would be thrown

This is a step to take before being able to disable registration forms
but still allow OAuth to register users.