Add reusable build information card (#2543)
Most build-specific pages currently display a basic table of information
about that build. For example:

![image](https://github.com/user-attachments/assets/11743b47-d581-41a4-8a48-42d26afb32e6)

This primitive approach uses a significant amount of vertical space,
while wasting most of the horizontal space available at the top of the
page. Additionally, this table does not display site information or
build history information, which means that it is often augmented by UI
elements which take up additional vertical space.

This PR introduces a new reusable "build information card" Vue component
which accepts a build ID and displays a large amount of summary
information about a build and associated site in a collapsible format.

Example of the card in its closed form:

![image](https://github.com/user-attachments/assets/ce207920-b0e8-4329-b503-93db28ff6e3a)

Example of the same card in its open form:

![image](https://github.com/user-attachments/assets/953907e3-1c8a-43f3-b0cf-38524a04f0db)

This PR implements the basic functionality, but there are a few future
improvements to be made, including:
- The horizontal space to the right of the build summary card is
reserved for charts showing the history of the build.
- The horizontal space to the right of the site information card is
reserved for charts showing the time allocation of the site, including
the amount of time spent on the associated project. A map showing the
location of the site (if site locations are enabled) is another future
improvement to be made.

I have added this component to the list of components to be tested via
Cypress component testing once type-safe GraphQL mocking is ready for
use.