Add "claim site" functionality to site overview page (#2680)
As discussed in #2659, the "claim site" feature is currently broken.
This PR supersedes the previous functionality by adding a "claim site"
button to the site overview page. I intend to make a separate PR to add
similar buttons to the project sites page. In addition, I plan to remove
`editSite.php` and move the remaining functionality to the site overview
page in a separate upcoming PR. That PR will also remove the broken
claim site logic.


![image](https://github.com/user-attachments/assets/f214bef4-0851-4c7d-a870-4fac6f6a2d00)



Closes #2659.