Wait for items to be displayed in SitesIdPageTest (#2696)
The new browser test `SitesIdPageTest` is somewhat flaky in CI because
it does not wait for the elements being checked to be visible. This is
the major downside to using Laravel's Dusk Selenium wrapper instead of
Cypress, which automatically waits for elements to be visible before all
assertions. This PR will hopefully resolve the flakiness by waiting for
the site history column to be displayed before making assertions.