Make buildgroup description column nullable (#2704)
The `buildgroup` table's `description` column is not-null for systems
created since the switch to Laravel, but this column may be nullable for
older systems. Given that the description is not a mandatory field, it
makes sense to make this field nullable for all systems. For older
systems, nothing will change. For newer systems, the buildgroup
description will now be nullable.