Standardize header menus (#2706)
The header menus are created differently on AngularJS, Vue, and XSL
pages. Since project information is now available everywhere at page
load time, we can render that information immediately instead of making
a separate request to fetch it. The upside of this is that users will
see the content immediately.