Introduce try/catch for image creation
As reported in #2682, invalid images (but valid a valid file) would
report a 500 error when CDash attempts to create the image from the
information.  Catch the error on creation and return early, as if the
image had an unknown extension.

Set the checksum to 0 to avoid writing other error messages to the log