Add XML path matching method to XML handlers (#2712)
This PR pulls out some of the logic in
https://github.com/Kitware/CDash/pull/2612 for matching full XML paths
into a separate PR in anticipation of several upcoming PRs. By
specifying the full XML path for an element, the exact element being
targeted is more clear. For the sake of example, I used the new path
method to locate all of the `<Site>` elements.