Remove broken `autoremovebuilds` test (#2784)
The `autoremovebuilds` test is completely broken, and is not hooked up
to anything in our test infrastructure. Given that the functionality it
tests has been superseded by newer logic, this test is no longer
relevant. This PR pulls out these specific changes from the broader
changes in #2698 to reduce the number of different changes in #2698.